refactor: 피드백 반영 className 적용
