+++ src/components/modal/payStub/AddPayStub/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const AddPaySub = () => {
+  return <div>index</div>;
+};
+
+export default AddPaySub;
