+++ src/components/modal/payStub/NomalPayStub/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const NomalPayStub = () => {
+  return <div>index</div>;
+};
+
+export default NomalPayStub;
