+++ src/pages/user/SalaryDetailsPage/index.tsx
