+++ src/pages/user/index.ts
@@ -2,5 +2,5 @@ export { default as UserHome } from './UserHome';
 export { default as Schedule } from './Schedule';
 export { default as EditProfile } from './EditProfile';
 export { default as Attendance } from './Attendance';
-export { default as SalaryCorrection } from './SalaryCorrection';
-export { default as SalaryDetails } from './SalaryDetails';
+export { default as SalaryCorrection } from './SalaryCorrectionPage';
+export { default as SalaryDetails } from './SalaryDetailsPage';
